\name{excel}

\alias{excel}

\title{
Save Excel Output
}

\description{
Writes processed data and methodological summaries to Excel file.
}

\usage{
excel(files, data)
}

\arguments{
\item{files}{list. As created by \code{\link{check_files}}.}
\item{data}{list. As created by \code{\link{process_data}}}
}
