\name{file_lookup}

\title{
Lookup Data Files
}

\description{
Finds CRDS and Sample ID files for specified instrument and date.
}

\usage{
file_lookup(instrument, runDate, dataPath)
}

\arguments{
\item{instrument}{character. Instrument serial number, e.g., \dQuote{hids2052}.}
\item{runDate}{character. Run start date, as character string in format
\dQuote{\%Y-\%m-\%d}}
\item{dataPath}{character. File path to folder containing CRDS data files.}
}

\examples{

}
